requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
matplotlib==3.8.1
pandas==2.1.3
Pillow==10.1.0
APScheduler==3.10.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9  # Для PostgreSQL
openpyxl==3.1.5        # Для работы с Excel файлами